+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.0.6 (In development)
+* Move diagnostic executor to seperate thread
+* Migrate to diag_server_unified ecu_diagnostics branch
+* Show diagnostic mode on status bar
+* Chart rendering at 60fps - Along with linear interpolation!
+* Initial SCN Configuration wizard:
+    1. Save/Load from YML files
+    2. Write/Read program settings
+    3. Wiki integration
+* Add support for ResponsePending ECU response (useful when flashing)
+
+
 # 1.0.3 (27/2/23)
 * Removed "UNDER DEVELOPMENT" watermark for V1.3 PCB